+++ src/form.tsx
@@ -1,11 +1,8 @@
 import * as React from 'react'
-import { formService, submissionService } from '@oneblink/apps'
-import { getRequest } from '@oneblink/apps/dist/services/fetch'
-import tenants from '@oneblink/apps/dist/tenants'
+import { formService, submissionService, formsAppService } from '@oneblink/apps'
 import { OneBlinkForm } from '@oneblink/apps-react'
 import OnLoading from '@oneblink/apps-react/dist/components/renderer/OnLoading'
-import { BrowserRouter as Router } from 'react-router-dom'
-import { FormsAppsTypes } from '@oneblink/types'
+import { useHistory } from 'react-router-dom'
 import ErrorModal from './ErrorModal'
 
 type Props = {
@@ -17,18 +14,11 @@ type Props = {
   externalId?: string
   googleMapsApiKey?: string
   abnLookupAuthenticationGuid?: string
+  paymentReceiptUrl?: string
 }
 
 const formIsSubmittingContainerStyles: React.CSSProperties = {
   opacity: 0.7,
-}
-
-async function getFormsAppConfiguration(formsAppId: number) {
-  const url = `${tenants.current.apiOrigin}/forms-apps/${formsAppId}/hostname-configuration`
-  const formsAppConfiguration = await getRequest<
-    FormsAppsTypes.FormsAppConfiguration<FormsAppsTypes.BaseFormsAppStyles>
-  >(url)
-  return formsAppConfiguration
 }
 
 function Form({
@@ -36,11 +26,13 @@ function Form({
   formsAppId,
   submissionRedirectUrl,
   cancelRedirectUrl,
+  paymentReceiptUrl,
   preFillData,
   externalId,
   googleMapsApiKey,
   abnLookupAuthenticationGuid,
 }: Props) {
+  const history = useHistory()
   const [
     { isFetching, form, formsAppConfiguration, fetchError },
     setFetchingState,
@@ -88,8 +80,15 @@ function Form({
 
         const formSubmissionResult = await submissionService.submit({
           formSubmission,
-          paymentReceiptUrl: null,
+          paymentReceiptUrl: paymentReceiptUrl,
         })
+        if (formSubmissionResult.submissionId && formSubmissionResult.payment) {
+          return submissionService.executePostSubmissionAction(
+            formSubmissionResult,
+            history.push,
+          )
+        }
+
         const url = new URL(submissionRedirectUrl)
         url.searchParams.append(
           'submissionId',
@@ -105,7 +104,13 @@ function Form({
         })
       }
     },
-    [externalId, formsAppId, submissionRedirectUrl],
+    [
+      externalId,
+      formsAppId,
+      submissionRedirectUrl,
+      paymentReceiptUrl,
+      history.push,
+    ],
   )
 
   const handleCancel = React.useCallback(() => {
@@ -128,7 +133,7 @@ function Form({
       try {
         const [f, formsAppConfiguration] = await Promise.all([
           formService.getForm(formId),
-          getFormsAppConfiguration(formsAppId),
+          formsAppService.getFormsAppConfiguration(formsAppId),
         ])
         setFetchingState({
           isFetching: false,
@@ -177,8 +182,7 @@ function Form({
   }
 
   return (
-    // apps-react won't render a form and instead throws an error unless wrapped in a router tag
-    <Router>
+    <>
       <div style={isSubmitting ? formIsSubmittingContainerStyles : undefined}>
         <OneBlinkForm
           form={form}
@@ -193,7 +197,7 @@ function Form({
         />
       </div>
       <ErrorModal error={submitError} onClose={clearSubmitError} />
-    </Router>
+    </>
   )
 }